* feat(workflow-npm): 二进制内置 read/list 子命令 + 工程化拆分 + e2e 测试

将独立脚本 scripts/read-run.ts 的读取能力合一进 peri-workflow 二进制，
新增 CLI 子命令模式（read/list/--help），无参数时保持 JSON-RPC 模式不变。

工程化：
- 主文件拆分为 src/ 下七个模块（types/rpc/adapter/server/reader/cli/index）
- 构建补充：bun build 打包单文件，tsc --noEmit 类型检查
- 测试补充：6 个测试文件 52 个用例，含 e2e 黑盒模拟（宿主 ↔ 真实二进制，
  覆盖全链路/未知方法/kill/CLI 错误路径），行覆盖率 94.39% ≥ 80% 门禁
- scripts/check-coverage.mjs 解析 bun 覆盖率表格（| 分隔列 + 百分比数值）


* feat(workflow-npm): 新增 validate 子命令校验 workflow 脚本语法

agent 编写 workflow 脚本时常犯三个错，引擎 parseScript 查不出、只能在运行时炸：workflow.agent(...) 旧式调用、缺 export const meta、无顶层 return。新增 peri-workflow validate <script.mjs> [--json] 子命令，在引擎检查（语法/export/import/meta 字面量）之上补充静态检查：

- error: 旧式 workflow.xxx( 调用（附修复指引）
- error: 缺 export const meta（区分真缺失与 meta 非法两类 case）
- warning: 无 return 语句（脚本将返回 undefined）

测试：validate 纯函数 11 例 + cli 分支 6 例 + e2e 黑盒 3 例，总计 72 tests / 0 fail，行覆盖率 94.64% ≥ 80%。真实冒烟：仓库 24 个脚本中 22 个通过、2 个旧格式脚本被正确标错。


* fix(statusbar): 修复多轮运行后模型快速切换弹窗点击失效

折行点落在模型段之后时，旧 span 粒度点击区域用最后一个 span 的行号
判定点击（模型文本在 row0 而判定用 row1），弹窗永久无法打开。

- model_click_areas 重写为词级折行模拟（对齐 ratatui-widgets 0.3.2
  reflow.rs Wrap{trim:false} 语义）：词流扫描 + 每词一个点击区域，
  模型段跨行/尾部折行时每行各有区域
- 换行条件用逐字符增量检查的精确等价形式
  line_x + ws + w - cw_last >= area_w（修复词尾 CJK 超界 1 列时
  的错误换行导致的点击失效残留），含行尾空白回填丢弃语义
- 测试矩阵扩充至 28 个：4 个 TestBackend ground-truth 渲染对比
  （逐位断言词首/词末字符位置）+ 等号边界/CJK 宽度/跨 span 词合并/
  词尾 CJK 超界场景；完整套件 694 测试通过，clippy 零警告
- 附带：issue 文档更新修复记录（含独立缺陷登记）


* feat(statusbar): 模型快速切换弹窗支持点击外部关闭

点击弹窗矩形（含全边框）之外任意位置即关闭弹窗（dismiss-on-outside-click），
与 Esc 关闭语义并列。点击边框/padding 不关闭、Consumed 防事件穿透到背景。

- 新增纯函数 click_inside_popup（矩形含边框判定，可测试）
- 鼠标 Down(Left) 分支：内容行切换并关闭（原行为不变）→ 矩形外 close_popup
- 新增 test_click_inside_popup（矩形内 6 点 + 四方向矩形外断言）


* perf(tui): 修复多 agent 并发 CPU 占用 50% 的三项核心瓶颈

静态诊断 + 双轮对抗验证（spec/issues/2026-08-02-multi-agent-concurrent-cpu-high.md），
三项中风险主因各派 coder 修复：

- entry.rs: 心跳 50ms 无条件写 → 100ms tick + 帧序号条件写。
  spinner 动画帧率本质 10Hz（elapsed_ms/100），50ms 是 2 倍超采样，
  一半重绘为重复帧；loading 期渲染下限 20Hz → 10Hz，动画零损失
- acp_types/tui_render_unit/render: 每 token 全量 VM 重建 → 增量修补。
  滚动哈希原语（增量与全量值一致）、sync_cache 只修补变化部分、
  删除 child_turn 深拷贝 + String 拼接 hash 链（im::Vector 共享 + u64 组合），
  每 token 从 O(N²) 降至 O(变化量+段数)
- markdown: 每 token 全量解析 → 尾部不稳定块回滚 + 持久化条件放宽。
  ensure_closed_code_fences 返回 Cow 零拷贝、缓存命中 mem::take、
  stable_text 增量扩展；流式散文不再缓存恒失效（3 轮解析器探针验证正确性）

验证：cargo build --workspace 通过；peri-tui 704 测试通过（含 9 个新增
markdown 回归测试）；clippy -D warnings 通过


* fix(tui): 消息区滚动与文本选区 usize 化，修复超长内容滚动不到底与选区错位

- ScrollPos 替代 ScrollViewState：offset 从 u16 升 usize，解除 65535 视觉行上限，
  scroll_to_bottom 设 usize::MAX 由渲染每帧 clamp 到 max_scroll
- 键盘翻页按当前视口高度（修复 PageUp/PageDown 只滚 1 行）
- resize 缩小视口后底部 footer/spinner 消失：vis_height 加入 effect 依赖 +
  prev_vis_height 哨兵 + should_follow_after_resize（resize 前在底部附近才跟随）
- 文本选区视觉行坐标 u16→usize：TextSelection/extract_visual_range/
  highlight_line_in_selection/visual_to_logical，超长内容拖拽选中不再错位
- 新增 ScrollPos/resize 跟随回归测试（含 >65535 行滚动测试）


* feat(tui): 消息区 footer 新增 keepgoing 按钮，点击继续 agent loop

空闲态（summary 行）右侧显示 [Keep Going] 按钮，点击发送空白 user prompt：
- peri-agent: append_messages_to_transcript 跳过空 Prompt——loop 因 consumed>0
  继续，但空消息不进 transcript/LLM 上下文
- peri-acp: run_session_loop 检测空白 content → 清空 incoming_recalls
- peri-tui: SubmitRequest::KeepGoing 变体 + handle_keepgoing_submit（无 session
  no-op、不产生本地气泡）；footer 按钮 span（防抖 muted 样式）+ 屏幕点击区域
  compute_keepgoing_rect（滚动跟随/视口裁剪）；click handler 注册在 scroll
  handler 之前（同优先级按注册序分发，避免被选区逻辑截断）；点击后 1.5s 防抖
  + tokio 恢复任务；i18n msg-keepgoing（en/zh-CN）
- 测试：空 Prompt 跳过 transcript、compute_keepgoing_rect 5 例


* fix(tui): keepgoing 按钮收尾——拼写修复、窄终端隐藏、防抖语义与判空统一

- atoms: KEEPGONG_BLOCKED_UNTIL → KEEPGOING_BLOCKED_UNTIL（拼写修复）
- footer: 窄终端下 summary+按钮超宽时跳过按钮渲染，避免换行后点击区域错位（m4）
- mod.rs: 防抖期内点击禁用按钮也 Consumed，事件不再落到文本选区逻辑；
  vis_width 前移供 footer 预计算使用
- executor: is_keepgoing 按 content block 判空（纯附件消息不误判）；
  空历史+空 prompt 直接短路，避免无意义 LLM 调用
- stages: append_messages_to_transcript 判空语义与 peri-acp is_keepgoing 统一


* fix(agent): 修复 keepgoing 空 Prompt 判空测试与 is_keepgoing 语义不一致

b83a0701 将 append_messages_to_transcript 判空统一为 content-block 判空
（与 peri-acp is_keepgoing 一致，纯附件消息不误判），但 1086ffa5 引入的
测试仍用空白文本断言跳过，导致回归。

- test_append_messages_empty_prompt_skipped 改用真实 keepgoing payload
  MessageContent::text("")（与 TUI handle_keepgoing_submit 一致）
- 新增 test_append_messages_whitespace_prompt_kept 锁定新语义：空白文本
  正常写入，防止改回 trim 判空破坏与 is_keepgoing 的一致性


---------

+++ peri-tui/src/kit/message_area/selection.rs
@@ -70,8 +70,8 @@ pub(super) fn concat_wrap_maps(
 }
 
 /// 二分查找：视觉行 → 逻辑行索引。
-pub(super) fn visual_to_logical(visual_row: u16, wrap_map: &[WrappedLineInfo]) -> Option<usize> {
-    let vr = visual_row as usize;
+pub(super) fn visual_to_logical(visual_row: usize, wrap_map: &[WrappedLineInfo]) -> Option<usize> {
+    let vr = visual_row;
     match wrap_map.binary_search_by(|entry| {
         if vr < entry.visual_start {
             Ordering::Greater
@@ -235,12 +235,13 @@ fn row_end_byte(plain: &str, row_start_byte: usize, col_in_row: u16) -> usize {
 ///
 /// [Why] 旧实现是整逻辑行高亮（粗粒度），与字符级复制提取不一致——用户看到的高亮
 /// 范围比实际复制内容大。改为字符级后高亮范围 = 复制范围，与终端鼠标拖拽选择一致。
+/// sel_sr/sel_er 为视觉行（usize，内容可超 65535 视觉行），sel_sc/sel_ec 为视觉列（u16）。
 #[allow(clippy::too_many_arguments)]
 pub(super) fn highlight_line_in_selection(
     line: &Line<'static>,
     entry: &WrappedLineInfo,
-    sel_sr: u16,
-    sel_er: u16,
+    sel_sr: usize,
+    sel_er: usize,
     sel_sc: u16,
     sel_ec: u16,
     width: u16,
@@ -250,18 +251,12 @@ pub(super) fn highlight_line_in_selection(
     let row_starts = wrap_byte_starts(line, &plain, width);
     let row_max = row_starts.len().saturating_sub(1);
 
-    let sr_in_line =
-        (sel_sr as usize) >= entry.visual_start && (sel_sr as usize) < entry.visual_end;
-    let er_in_line =
-        (sel_er as usize) >= entry.visual_start && (sel_er as usize) < entry.visual_end;
+    let sr_in_line = sel_sr >= entry.visual_start && sel_sr < entry.visual_end;
+    let er_in_line = sel_er >= entry.visual_start && sel_er < entry.visual_end;
 
     // sr_off / er_off 是选区起点/终点视觉行相对该逻辑行 visual_start 的偏移
-    let sr_off = (sel_sr as usize)
-        .saturating_sub(entry.visual_start)
-        .min(row_max);
-    let er_off = (sel_er as usize)
-        .saturating_sub(entry.visual_start)
-        .min(row_max);
+    let sr_off = sel_sr.saturating_sub(entry.visual_start).min(row_max);
+    let er_off = sel_er.saturating_sub(entry.visual_start).min(row_max);
 
     let (b_start, b_end) = if sr_in_line && er_in_line {
         // 同一行：byte 范围 = [sr_off 行内 sc, er_off 行内 ec]
@@ -368,12 +363,13 @@ pub(super) fn mark_copy_message(char_count: usize) {
 /// [TRAP] 选区可能超出 core 范围（footer 区域无 wrap_map）——clamp 到 wrap_map
 /// 末尾，确保 footer 行的 visual_to_logical 不返回 None 导致整个提取失败。
 /// [Scheme D] 通过 slot_arcs + slot_offsets 按需从 slot 中解析行，不再依赖全量 lines 切片。
+/// vis_start/vis_end 为视觉坐标：行 usize（内容可超 65535 视觉行）、列 u16。
 pub(super) fn extract_visual_range(
     slots: &[Arc<Vec<Line<'static>>>],
     slot_offsets: &[usize],
     wrap_map: &[WrappedLineInfo],
-    vis_start: (u16, u16),
-    vis_end: (u16, u16),
+    vis_start: (usize, u16),
+    vis_end: (usize, u16),
     width: u16,
 ) -> Option<String> {
     let ((sr, sc), (er, ec)) = if vis_start <= vis_end {
@@ -384,7 +380,7 @@ pub(super) fn extract_visual_range(
     // Clamp sr/er 到 wrap_map 视觉范围内（footer 区域无 wrap_map，避免 None）
     let max_visual = wrap_map
         .last()
-        .map(|e| (e.visual_end.saturating_sub(1)) as u16)
+        .map(|e| e.visual_end.saturating_sub(1))
         .unwrap_or(0);
     let sr = sr.min(max_visual);
     let er = er.min(max_visual);
@@ -403,12 +399,8 @@ pub(super) fn extract_visual_range(
         let row_starts = wrap_byte_starts(line, &plain, width);
         // 把视觉行号 clamp 到 row_starts 索引范围内（防御：footer 区域等异常 sr/er）
         let row_max = row_starts.len().saturating_sub(1);
-        let sr_off = (sr as usize)
-            .saturating_sub(entry.visual_start)
-            .min(row_max);
-        let er_off = (er as usize)
-            .saturating_sub(entry.visual_start)
-            .min(row_max);
+        let sr_off = sr.saturating_sub(entry.visual_start).min(row_max);
+        let er_off = er.saturating_sub(entry.visual_start).min(row_max);
 
         if first == last {
             // 同一逻辑行：起点行 sr_off 内的列 sc → 终点行 er_off 内的列 ec
